requirements.txt
ta
pandas
python-telegram-bot
requests
